chore(task4): remove unused code

+++ task04/city/city.cpp
@@ -32,7 +32,7 @@ namespace tz {
         return this->position.getY();
     }
 
-    unsigned long City::getNumberOfPOIs() const {
+    int City::getNumberOfPOIs() const {
         return this->currentSize;
     }
